mocks refactoring and types fix
